imx95-a1-19x19-verdin.conf: Add new machine for A1 support
Soon, the B0 SoC revision will take over as the default for the i.MX95.

Add a new machine for pre-production A1 support. This machine can be
used for A0/A1 SoC revisions when the imx95-19x19-verdin.conf gets
updated to support the B0 or later SoC revisions.

Add the "imx95-19x19-verdin" machine override such that all
conditional overrides for the base machine are also executed for this A1
machine.

In practice this override was only used in `imx-boot_1.0.bb` for
MKIMAGE_EXTRA_ARGS. meta-imx added a `imx95-a1-19x19-verdin` override
there that does the same as the `imx95-19x19-verdin` override, this
commit solves it differently by adding the `imx95-19x19-verdin` override
to the A1 machine.

Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>